+++ src/components/Icons/Alarm.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconsProps } from './IconsProps';
+import type { IconsProps } from './dto';
 
 const Alarm: React.FC<IconsProps> = ({ isActive = false }) => {
 	return (
