+++ src/components/Icons/Heart.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconsProps } from './IconsProps';
+import type { IconsProps } from './dto';
 
 const Heart: React.FC<IconsProps> = ({ isFilled = false }) => {
 	return (
