+++ src/components/Icons/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconsProps } from './IconsProps';
+import type { IconsProps } from './dto';
 
 const Home: React.FC<IconsProps> = ({ color = '', isFilled = false }) => {
 	return (
