+++ src/components/Icons/Like.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconsProps } from './IconsProps';
+import type { IconsProps } from './dto';
 
 const Like: React.FC<IconsProps> = ({ isFilled = false, color = '' }) => {
 	if (color && !isFilled) {
